refactor: extract SaisieContext
